refactor: separate platform code into architectural planes
Extract control-plane, application-plane, and data-plane boundaries:

- Create TALXIS.CLI.Platform.PowerPlatform.Control for Power Platform
  admin API / environment catalog (control plane)
- Create TALXIS.CLI.Platform.Dataverse.Application for solution, package,
  and deployment services + platform helpers (application plane)
- Retain TALXIS.CLI.Platform.Dataverse for runtime, auth, MSAL,
  connection factory, and bootstrapping (data plane)
- Add provider-agnostic IAccessTokenService and IHttpClientFactoryWrapper
  abstractions to Core so control-plane code does not depend on Dataverse
- Split DI registration: runtime services in AddTxcDataverseProvider(),
  application services in AddTxcDataverseApplication()
- Add docs/architecture.md documenting plane boundaries and contributor
  guidance for new code placement

All 413 unit tests and 26 integration tests pass. External package IDs
and command aliases (txc, txc-mcp) are unchanged.
