+++ TALXIS.CLI.sln
@@ -35,6 +35,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TALXIS.CLI.Core", "src\TALX
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TALXIS.CLI.Features.Config", "src\TALXIS.CLI.Features.Config\TALXIS.CLI.Features.Config.csproj", "{5318996C-5E12-4C6B-9344-883CD9E2F4B7}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TALXIS.CLI.Platform.PowerPlatform.Control", "src\TALXIS.CLI.Platform.PowerPlatform.Control\TALXIS.CLI.Platform.PowerPlatform.Control.csproj", "{CE13B13C-FC46-4318-BD46-3E44C97E1554}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TALXIS.CLI.Platform.Dataverse.Application", "src\TALXIS.CLI.Platform.Dataverse.Application\TALXIS.CLI.Platform.Dataverse.Application.csproj", "{55DEF0A8-193B-4B7E-A646-0D3911FABB88}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -213,6 +217,30 @@ Global
 		{5318996C-5E12-4C6B-9344-883CD9E2F4B7}.Release|x64.Build.0 = Release|Any CPU
 		{5318996C-5E12-4C6B-9344-883CD9E2F4B7}.Release|x86.ActiveCfg = Release|Any CPU
 		{5318996C-5E12-4C6B-9344-883CD9E2F4B7}.Release|x86.Build.0 = Release|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Debug|x64.Build.0 = Debug|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Debug|x86.Build.0 = Debug|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Release|Any CPU.Build.0 = Release|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Release|x64.ActiveCfg = Release|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Release|x64.Build.0 = Release|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Release|x86.ActiveCfg = Release|Any CPU
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554}.Release|x86.Build.0 = Release|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Debug|x64.Build.0 = Debug|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Debug|x86.Build.0 = Debug|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Release|Any CPU.Build.0 = Release|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Release|x64.ActiveCfg = Release|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Release|x64.Build.0 = Release|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Release|x86.ActiveCfg = Release|Any CPU
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -232,6 +260,8 @@ Global
 		{ED036063-8EE3-43BB-8380-3A114CE515FD} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{80DCD8C5-8577-424F-B4E2-9A1651A069D6} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 		{5318996C-5E12-4C6B-9344-883CD9E2F4B7} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{CE13B13C-FC46-4318-BD46-3E44C97E1554} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{55DEF0A8-193B-4B7E-A646-0D3911FABB88} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {53733BD6-A32A-41B7-9472-E377AF68151F}
