+++ src/TALXIS.CLI.Platform.Dataverse.Application/Platforms/PackageHistoryWriter.cs
@@ -7,7 +7,7 @@ using Microsoft.Xrm.Sdk.Messages;
 using Microsoft.Xrm.Sdk.Metadata;
 using Microsoft.Xrm.Sdk.Query;
 
-namespace TALXIS.CLI.Platform.Dataverse.Platforms;
+namespace TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 
 public sealed record PackageHistoryStatusCodes(
     int? InProcessStatus,
