+++ src/TALXIS.CLI.Platform.Dataverse.Application/Platforms/SolutionImporter.cs
@@ -8,9 +8,10 @@ using Microsoft.Xrm.Sdk.Messages;
 using Microsoft.Xrm.Sdk.Query;
 using TALXIS.CLI.Core.Platforms.Dataverse;
 using TALXIS.CLI.Platform.Dataverse;
+using TALXIS.CLI.Platform.Dataverse.Application;
 using SolutionInfo = TALXIS.CLI.Core.Platforms.Dataverse.SolutionInfo;
 
-namespace TALXIS.CLI.Platform.Dataverse.Platforms;
+namespace TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 
 /// <summary>
 /// Orchestrates solution imports against Dataverse via the modern <see cref="ServiceClient"/>.
