+++ src/TALXIS.CLI.Platform.Dataverse.Application/Services/DataverseDataPackageService.cs
@@ -1,11 +1,11 @@
 using Microsoft.Identity.Client;
 using TALXIS.CLI.Core.Abstractions;
 using TALXIS.CLI.Core.Platforms.Dataverse;
-using TALXIS.CLI.Platform.Dataverse.Platforms;
+using TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 using TALXIS.CLI.Platform.Dataverse.Runtime;
 using TALXIS.CLI.Platform.Xrm;
 
-namespace TALXIS.CLI.Platform.Dataverse.Services;
+namespace TALXIS.CLI.Platform.Dataverse.Application.Services;
 
 internal sealed class DataverseDataPackageService : IDataPackageService
 {
