+++ src/TALXIS.CLI.Platform.Dataverse.Application/Services/DataversePackageImportService.cs
@@ -1,10 +1,10 @@
 using Microsoft.Identity.Client;
 using TALXIS.CLI.Core.Platforms.Dataverse;
-using TALXIS.CLI.Platform.Dataverse.Platforms;
+using TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 using TALXIS.CLI.Platform.Dataverse.Runtime;
 using TALXIS.CLI.Platform.Xrm;
 
-namespace TALXIS.CLI.Platform.Dataverse.Services;
+namespace TALXIS.CLI.Platform.Dataverse.Application.Services;
 
 internal sealed class DataversePackageImportService : IPackageImportService
 {
