+++ src/TALXIS.CLI.Platform.Dataverse.Application/Services/DataverseSolutionUninstallService.cs
@@ -1,10 +1,10 @@
 using Microsoft.Extensions.Logging;
 using TALXIS.CLI.Core.Platforms.Dataverse;
-using TALXIS.CLI.Platform.Dataverse.Platforms;
+using TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 using TALXIS.CLI.Platform.Dataverse.Runtime;
 using TALXIS.CLI.Logging;
 
-namespace TALXIS.CLI.Platform.Dataverse.Services;
+namespace TALXIS.CLI.Platform.Dataverse.Application.Services;
 
 internal sealed class DataverseSolutionUninstallService : ISolutionUninstallService
 {
