+++ src/TALXIS.CLI.Platform.Dataverse.Application/TALXIS.CLI.Platform.Dataverse.Application.csproj
@@ -4,21 +4,17 @@
     <TargetFramework>net10.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <RootNamespace>TALXIS.CLI.Platform.Dataverse</RootNamespace>
+    <RootNamespace>TALXIS.CLI.Platform.Dataverse.Application</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.DependencyInjection.Abstractions" Version="10.0.5" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="10.0.5" />
-    <PackageReference Include="Microsoft.Identity.Client" Version="4.83.3" />
-    <PackageReference Include="Microsoft.Identity.Client.Extensions.Msal" Version="4.83.3" />
     <PackageReference Include="Microsoft.PowerPlatform.Dataverse.Client" Version="1.2.10" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\TALXIS.CLI.Core\TALXIS.CLI.Core.csproj" />
     <ProjectReference Include="..\TALXIS.CLI.Logging\TALXIS.CLI.Logging.csproj" />
-    <ProjectReference Include="..\TALXIS.CLI.Platform.PowerPlatform.Control\TALXIS.CLI.Platform.PowerPlatform.Control.csproj" />
+    <ProjectReference Include="..\TALXIS.CLI.Platform.Dataverse\TALXIS.CLI.Platform.Dataverse.csproj" />
     <ProjectReference Include="..\TALXIS.CLI.Platform.Xrm\TALXIS.CLI.Platform.Xrm.csproj" />
   </ItemGroup>
