+++ src/TALXIS.CLI.Platform.Dataverse/DependencyInjection/DataverseProviderServiceCollectionExtensions.cs
@@ -1,12 +1,10 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using TALXIS.CLI.Core.Abstractions;
-using TALXIS.CLI.Core.Platforms.Dataverse;
 using TALXIS.CLI.Platform.Dataverse.Authority;
 using TALXIS.CLI.Platform.Dataverse.Runtime;
 using TALXIS.CLI.Platform.Dataverse.Msal;
-using TALXIS.CLI.Platform.Dataverse.PowerPlatform;
-using TALXIS.CLI.Platform.Dataverse.Services;
+using TALXIS.CLI.Platform.PowerPlatform.Control;
 using TALXIS.CLI.Core.Resolution;
 using TALXIS.CLI.Core.Storage;
 
@@ -15,11 +13,18 @@ namespace TALXIS.CLI.Platform.Dataverse.DependencyInjection;
 public static class DataverseProviderServiceCollectionExtensions
 {
     /// <summary>
-    /// Registers the Dataverse <see cref="IConnectionProvider"/>, the shared
-    /// MSAL client factory, the authority-challenge resolver, and the MSAL
-    /// token-cache binder. Must be called after
-    /// <c>AddTxcConfigCore()</c>.
+    /// Registers the Dataverse runtime infrastructure: MSAL client factory,
+    /// authority-challenge resolver, token-cache binder, connection provider,
+    /// access-token service, connection factory, live checker, interactive
+    /// login, bootstrapper, and the Power Platform environment catalog.
+    /// Must be called after <c>AddTxcConfigCore()</c>.
     /// </summary>
+    /// <remarks>
+    /// Application-plane service implementations (solution import, deployment
+    /// history, etc.) are registered separately via
+    /// <c>AddTxcDataverseApplicationServices()</c> in
+    /// <see cref="TALXIS.CLI.Platform.Dataverse.Application"/> .
+    /// </remarks>
     public static IServiceCollection AddTxcDataverseProvider(this IServiceCollection services)
     {
         services.AddSingleton<DataverseMsalClientFactory>();
@@ -39,19 +44,13 @@ public static class DataverseProviderServiceCollectionExtensions
         services.AddSingleton<ITokenCacheStore>(sp => sp.GetRequiredService<DataverseTokenCacheBinder>());
 
         services.AddSingleton<IConnectionProvider, DataverseConnectionProvider>();
-        services.AddSingleton<IDataverseAccessTokenService, DataverseAccessTokenService>();
+        services.AddSingleton<DataverseAccessTokenService>();
+        services.AddSingleton<IDataverseAccessTokenService>(sp => sp.GetRequiredService<DataverseAccessTokenService>());
+        services.AddSingleton<IAccessTokenService>(sp => sp.GetRequiredService<DataverseAccessTokenService>());
         services.AddSingleton<IDataverseConnectionFactory, DataverseConnectionFactory>();
         services.AddSingleton<IPowerPlatformEnvironmentCatalog, PowerPlatformEnvironmentCatalog>();
         services.AddSingleton<IDataverseLiveChecker, DataverseLiveChecker>();
         services.AddSingleton<IInteractiveLoginService, DataverseInteractiveLoginService>();
-        services.AddSingleton<ISolutionInventoryService, DataverseSolutionInventoryService>();
-        services.AddSingleton<IDataPackageService, DataverseDataPackageService>();
-        services.AddSingleton<ISolutionUninstallService, DataverseSolutionUninstallService>();
-        services.AddSingleton<ISolutionImportService, DataverseSolutionImportService>();
-        services.AddSingleton<IDeploymentHistoryService, DataverseDeploymentHistoryService>();
-        services.AddSingleton<IDeploymentDetailService, DataverseDeploymentDetailService>();
-        services.AddSingleton<IPackageImportService, DataversePackageImportService>();
-        services.AddSingleton<IPackageUninstallService, DataversePackageUninstallService>();
         services.AddSingleton<TALXIS.CLI.Core.Bootstrapping.IConnectionProviderBootstrapper,
             TALXIS.CLI.Platform.Dataverse.Bootstrapping.DataverseConnectionProviderBootstrapper>();
         return services;
