+++ tests/TALXIS.CLI.Tests/Environment/Package/PackageUninstallCliCommandTests.cs
@@ -1,6 +1,6 @@
 using TALXIS.CLI.Features.Environment;
 using TALXIS.CLI.Features.Environment.Package;
-using TALXIS.CLI.Platform.Dataverse.Platforms;
+using TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 using Xunit;
 
 namespace TALXIS.CLI.Tests.Environment.Package;
