+++ tests/TALXIS.CLI.Tests/Environment/Platforms/Dataverse/DeploymentFindingsAnalyzerTests.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using TALXIS.CLI.Platform.Dataverse;
 using TALXIS.CLI.Features.Environment;
-using TALXIS.CLI.Platform.Dataverse.Platforms;
+using TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 using Xunit;
 
 namespace TALXIS.CLI.Tests.Environment.Platforms.Dataverse;
