+++ tests/TALXIS.CLI.Tests/Environment/Platforms/Dataverse/PackageHistoryWriterTests.cs
@@ -1,6 +1,6 @@
 using TALXIS.CLI.Platform.Dataverse;
 using TALXIS.CLI.Features.Environment;
-using TALXIS.CLI.Platform.Dataverse.Platforms;
+using TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 using Xunit;
 
 namespace TALXIS.CLI.Tests.Environment.Platforms.Dataverse;
