+++ tests/TALXIS.CLI.Tests/Environment/Platforms/Dataverse/SolutionHistoryReaderTests.cs
@@ -2,7 +2,7 @@ using System;
 using Microsoft.Xrm.Sdk;
 using TALXIS.CLI.Platform.Dataverse;
 using TALXIS.CLI.Features.Environment;
-using TALXIS.CLI.Platform.Dataverse.Platforms;
+using TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 using Xunit;
 
 namespace TALXIS.CLI.Tests.Environment.Platforms.Dataverse;
