+++ tests/TALXIS.CLI.Tests/Environment/Platforms/Dataverse/SolutionImporterPathSelectionTests.cs
@@ -2,7 +2,7 @@ using System.Xml.Linq;
 using TALXIS.CLI.Platform.Dataverse;
 using TALXIS.CLI.Features.Environment;
 using TALXIS.CLI.Core.Platforms.Dataverse;
-using TALXIS.CLI.Platform.Dataverse.Platforms;
+using TALXIS.CLI.Platform.Dataverse.Application.Platforms;
 using Xunit;
 
 namespace TALXIS.CLI.Tests.Environment.Platforms.Dataverse;
