+++ tests/TALXIS.CLI.Tests/TALXIS.CLI.Tests.csproj
@@ -20,6 +20,8 @@
     <ProjectReference Include="../../src/TALXIS.CLI.Logging/TALXIS.CLI.Logging.csproj" />
     <ProjectReference Include="../../src/TALXIS.CLI.MCP/TALXIS.CLI.MCP.csproj" />
     <ProjectReference Include="../../src/TALXIS.CLI.Platform.Dataverse/TALXIS.CLI.Platform.Dataverse.csproj" />
+    <ProjectReference Include="../../src/TALXIS.CLI.Platform.Dataverse.Application/TALXIS.CLI.Platform.Dataverse.Application.csproj" />
+    <ProjectReference Include="../../src/TALXIS.CLI.Platform.PowerPlatform.Control/TALXIS.CLI.Platform.PowerPlatform.Control.csproj" />
     <ProjectReference Include="../../src/TALXIS.CLI.Features.Environment/TALXIS.CLI.Features.Environment.csproj" />
     <ProjectReference Include="../../src/TALXIS.CLI.Features.Config/TALXIS.CLI.Features.Config.csproj" />
   </ItemGroup>
